Add files via upload
Simple Demo of JPanel in JPanel with proposed brutish button functionality toggling visibility. Will add to this to make strategy integration more understandable with current button functions. The idea is to keep everything in one GUI window. Its important to note, when a JPanelX is in JPanelY, setting JPanelY visibility to false makes JPanelX invisible also. A work around is to place all JPanels in same location instead of inside eachother. Otherwise, the component visibility of JPanelY should be toggled instead.