Reading bulk order from file has been added to CRUD
This application can now read text files specifically formatted for data input regarding sales order. For now, it parses the data and stores it locally, but after sql connection is coded in, we can have it update the database directly. The open file dialog is shown when the open file button is pressed in the crud screen. Please test it on your systems to make it sure it works butter smoothly so far