Update guiBtnSkeleton.java
Attempting to make LinkedList of Linkedlists. The database will be read into a linkedList, first with the unique product ID as the head of each node then the Product information as a linked list off each head node containing product ID. The productID linked list will be used in the search method and can then also be used in in the rest of the crud methods. The linked list will be able to search by product ID in the ProductID textfield entry box on the GUI. The other product info textfield boxes entry will render a search through each linkedlist's list of product information.